# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiEnvelope_ResponseLogin_": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseLogin"
              },
              {
                "type": "null"
              }
            ]
          },
          "internal_code": {
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ApiEnvelope[ResponseLogin]",
        "type": "object"
      },
      "ApiEnvelope_ResponseLogout_": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseLogout"
              },
              {
                "type": "null"
              }
            ]
          },
          "internal_code": {
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ApiEnvelope[ResponseLogout]",
        "type": "object"
      },
      "ApiEnvelope_ResponseMe_": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseMe"
              },
              {
                "type": "null"
              }
            ]
          },
          "internal_code": {
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ApiEnvelope[ResponseMe]",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "ErrorEnvelope_Literal_401__Literal__AUTH_INVALID_CREDENTIALS___": {
        "properties": {
          "code": {
            "const": 401,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "AUTH_INVALID_CREDENTIALS",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[401], Literal['AUTH_INVALID_CREDENTIALS']]",
        "type": "object"
      },
      "ErrorEnvelope_Literal_401__Literal__AUTH_SESSION_EXPIRED___": {
        "properties": {
          "code": {
            "const": 401,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "AUTH_SESSION_EXPIRED",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[401], Literal['AUTH_SESSION_EXPIRED']]",
        "type": "object"
      },
      "ErrorEnvelope_Literal_403__Literal__REQUEST_ORIGIN_INVALID___": {
        "properties": {
          "code": {
            "const": 403,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "REQUEST_ORIGIN_INVALID",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[403], Literal['REQUEST_ORIGIN_INVALID']]",
        "type": "object"
      },
...
      "ErrorEnvelope_Literal_422__Literal__REQUEST_INVALID___": {
        "properties": {
          "code": {
            "const": 422,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "REQUEST_INVALID",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[422], Literal['REQUEST_INVALID']]",
        "type": "object"
      },
      "ErrorEnvelope_Literal_500__Literal__INTERNAL_ERROR___": {
        "properties": {
          "code": {
            "const": 500,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "INTERNAL_ERROR",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[500], Literal['INTERNAL_ERROR']]",
        "type": "object"
      },
      "ErrorEnvelope_Literal_503__Literal__AUTH_STORE_INVALID___": {
        "properties": {
          "code": {
            "const": 503,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "AUTH_STORE_INVALID",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[503], Literal['AUTH_STORE_INVALID']]",
        "type": "object"
      },
      "ErrorEnvelope_Literal_503__Literal__AUTH_STORE_UNAVAILABLE___": {
        "properties": {
          "code": {
            "const": 503,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "internal_code": {
            "const": "AUTH_STORE_UNAVAILABLE",
            "title": "Internal Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "internal_code",
          "message",
          "data"
        ],
        "title": "ErrorEnvelope[Literal[503], Literal['AUTH_STORE_UNAVAILABLE']]",
        "type": "object"
      },
...
      "RequestLogin": {
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "RequestLogin",
        "type": "object"
      },
      "RequestLogout": {
        "properties": {},
        "title": "RequestLogout",
        "type": "object"
      },
      "RequestMe": {
        "properties": {},
        "title": "RequestMe",
        "type": "object"
      },
...
      "ResponseLogin": {
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "expiresIn": {
            "title": "Expiresin",
            "type": "integer"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "username",
          "displayName",
          "expiresIn"
        ],
        "title": "ResponseLogin",
        "type": "object"
      },
      "ResponseLogout": {
        "properties": {
          "loggedOut": {
            "title": "Loggedout",
            "type": "boolean"
          }
        },
        "required": [
          "loggedOut"
        ],
        "title": "ResponseLogout",
        "type": "object"
      },
      "ResponseMe": {
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "username",
          "displayName"
        ],
        "title": "ResponseMe",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Risk Workbench API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ResponseLogin_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_401__Literal__AUTH_INVALID_CREDENTIALS___"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_403__Literal__REQUEST_ORIGIN_INVALID___"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_422__Literal__REQUEST_INVALID___"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_500__Literal__INTERNAL_ERROR___"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__AUTH_STORE_UNAVAILABLE___"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestLogout"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ResponseLogout_"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_403__Literal__REQUEST_ORIGIN_INVALID___"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_422__Literal__REQUEST_INVALID___"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_500__Literal__INTERNAL_ERROR___"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__AUTH_STORE_UNAVAILABLE___"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "post": {
        "operationId": "me_api_v1_auth_me_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestMe"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ResponseMe_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_401__Literal__AUTH_SESSION_EXPIRED___"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_403__Literal__REQUEST_ORIGIN_INVALID___"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_422__Literal__REQUEST_INVALID___"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope_Literal_500__Literal__INTERNAL_ERROR___"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__AUTH_STORE_UNAVAILABLE___"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope_Literal_503__Literal__AUTH_STORE_INVALID___"
                    }
                  ],
                  "title": "Response 503 Me Api V1 Auth Me Post"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}